requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
SQLAlchemy[asyncio]>=2.0.28
asyncpg>=0.29.0
alembic>=1.13.1
celery>=5.3.6
redis>=5.0.1
fastapi-limiter>=0.1.6
jinja2>=3.1.3
httpx>=0.27.0
aioboto3>=12.4.0
structlog>=24.1.0
sentry-sdk>=1.40.6
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-exporter-otlp>=1.24.0
prometheus-fastapi-instrumentator>=6.1.0
python-multipart>=0.0.9
pytest>=8.2.0
